+++ objc2-foundation/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Changed
+* **BREAKING**: Renamed `INSFastEnumeration::enumerator` to
+  `INSFastEnumeration::iter_fast`.
+
 
 ## 0.2.0-alpha.3 - 2021-12-22